requirements.txt
gradio
matplotlib
numpy
opencv_contrib_python
opencv_python
opencv_python_headless
Pillow
tqdm
transformers
